split presenter
