+++ app/AccountancyModule/PaymentModule/templates/CreateEducationGroup/default.latte
@@ -1,3 +1,3 @@
 {block content}
 <h1 n:inner-block="title">Nová skupina plateb <small>{$education->displayName}</small></h1>
-{control newGroupForm}
+{control form}
