+++ app/AccountancyModule/PaymentModule/templates/GroupList/default.latte
@@ -52,7 +52,7 @@
 
                         <div class="dropdown-divider"></div>
 
-                        <a n:href="Education:default" class="dropdown-item">
+                        <a n:href="SelectEducationForGroup:" class="dropdown-item">
                             <i class="fas fa-user-graduate"></i> Vzdělávačka
                             <span class="small d-block">Umožnuje načíst osoby ze vzdělávací akce.</span>
                         </a>
