+++ app/AccountancyModule/PaymentModule/templates/SelectEducationForGroup/default.latte
@@ -18,7 +18,7 @@
                                     <h5 class="mb-1">{$education->displayName}</h5>
                                 </div>
                                 <div>
-                                    <a n:href="newGroup $education->id->toInt()" class="btn btn-outline-primary">Vytvořit skupinu</a>
+                                    <a n:href="CreateEducationGroup: $education->id->toInt()" class="btn btn-outline-primary">Vytvořit skupinu</a>
                                 </div>
                             </div>
                         </li>
